+++ StockAlertInfo.cs
@@ -4,6 +4,6 @@ namespace StockAlert
     {
         public const string Id = "com.stockalert.ats";
         public const string Name = "StockAlert";
-        public const string Version = "1.1.6";
+        public const string Version = "1.1.7";
     }
 }